adding monolith natural wondr -WIP
a natural wonder that gives an atrocious raider like promotion to units. i want natural wonders that encourage certain play-styles. i want salt the earth to be a natural wonder instead of a belief. also want all natural wonders to give a special promotion than discovers a an un-researchable tech.